Add colab notebook and rmd I used to get the csvs
The notebook contains the tabular deep learning model from fast.ai. The
rmd contains the code I used to create the csvs that I fed into the
model in the python notebook. Some of the code from the Rmd is
duplicated with the data cleaning rmd; I will migrate the code from this
file to the data cleaning rmd at some point.